==> part_02-data_structures.Rmd <==
---
title: "Data Types and Data Structures"
output: 
  github_document:
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

***

As in many programming languages, understanding how data are stored and manipulated is important. In this regard, the terms ‘data structure’ and ‘data type’ refer to how R treats your data rather than how you organize your data.

This introduction might become rather technical, but don't get upset. It is less important to remember everything since we will recapitulate the contents on a needs basis during the workshop.

## Data Types in R

Any object in R can be of _one single_ data type only. The most commonly encountered data types are shown below. 

| data type | stores | example |
|-----------|--------|---------|
| `double` | floating point numbers | `3.141, 1.8e-3` |
| `integer` | integers | `-5L, 1L, 2L, 3L` |
| `complex` | complex numbers | `1+2i, 3-1i` |
| `character` | strings | `"banana", "apple", "melon"` |
| `logical` | boolean values | `TRUE, FALSE` |
| `list` | sets with different data types | `list(3, "apple")` |
| `closure` | functions | `function(x) {return(x + 1)}` |

<!-- For the sake of completeness (but irrelevant for this workshop), everything that can be typed in the R console can be stored as an R object. Therefore, there are some more data types. -->

<!-- | data type | stores | example | -->
<!-- |-----------|--------|---------| -->
<!-- | `environment` | environments | `new.env()` | -->
<!-- | `language` | primitive expressions | `sapply(expression(x = 1 + 1, 2 + 3), typeof)` | -->
<!-- | `expression` | a set of (incoherent) expressions | `typeof(expression(x = 1 + 1, 2 + 3))` | -->

> By convention, if you type any number (including integers without an `L` suffix), R will store it as `double`. This allows to do accurate maths with it. (In computer science, ‘accurate’ means within the limits of the machine.) `double` stands for ‘double-precision floating-point format’.

The data type can be queried using the `typeof(x)` function. 

```{r}
x <- pi
typeof(x)
```

There are some special values, R objects can take, 

* `NULL` for empty objects; their data type is eradicated (`NULL`) too, 
* `NA`, which is used to designate a missing value; their data type is preserved, and
* `NaN` and `Inf` to designate non-numerical results of certain computations with `double` or `complex` objects.

## Atomic Data Structures: Vectors

In R, multiple objects of the same type can be _concatenated_ (or _combined_) to vectors of the same data type using the `c(...)` function. Or, to put it the other way round, in R, every object is at least a vector with length one.

```{r eval=FALSE}
x <- c(1, 2, 3)
typeof(x)
length(x)

x <- c(c(1, 2, 3), c(4, 5, 6))
typeof(x)
length(x)

x <- c("a", "b", "c")
typeof(x)
length(x)

x <- c(TRUE, FALSE, FALSE, TRUE)
typeof(x)
length(x)

x <- c(list("a", 1), list("b", 2), list("c", 3))
typeof(x)
length(x)
```

Elements in vectors can be _named_. 

```{r}
# names do not need to be quoted necessarily
noble_gases <- c(He = 2, "Ne" = 10, Ar = 18, Kr = 36, Xe = 54)
noble_gases

# get the names as character vector
names(noble_gases)

# set the names with another character vector
names(noble_gases) <- c("Helium", "Neon", "Argon", "Krypton", "Xenon")
noble_gases
```

We can add another entry by concatenation.

```{r}
noble_gases <- c(noble_gases, "Radon" = 86)
```

When we try to concatenate _objects of different data types_, the objects with the ‘least compatible’ data type will be **coerced** to an object of the ‘most compatible’ data type. The order being roughly `logical` < `integer` < `double` < `complex` < `character` < `list`.

```{r eval=FALSE}
# coercion to character
c(1, "a")
c(TRUE, "a")

# coercion to double
c(1L, 2.5)

# coercion to integer
x <- c(1L, TRUE, FALSE)
typeof(x)
```

Thus, to preserve the original data type of each element, we need to use lists!

We can also enforce R to coerce an appropriate value. If this is not possible, `NA` is returned.

```{r eval=FALSE}
as.double("1")
typeof(as.double("1"))  # double

as.double("a")
typeof(as.double("a"))  # NA of type double

as.logical("T")
typeof(as.logical("T")) # logical
```

## Advanced Data Structures: Classes

R allows to create more sophisticated data structures from the simpler atomic data types. The structure of such objectes is called a ‘class’. Besides its usefulness for an object-oriented programming style, having a class associated to an object will cause functions such as `print`, `plot` and `summary` to respond in an appropriate manner. Some functions accept only objects of specific classes.

In RStudio, roughly speaking, the class of an object is shown next to its name in the ‘Environment’ pane. Note that `integer`, `double` and `complex` are of class `numeric`.

> All objects in R can have additional “metadata” associated with them, so-called ‘attributes’. Classes define which ‘attributes’ an object must (at least) implement to belong to a class.

Here are two examples to illustrate the concept.

1. _‘Tables’_ (class `data.frame` or `data.table` or `tibble`) are objects of type `list`. 

    * They combine columns (each a vector with a single data type), so that different data types can be stored side-by-side. 
    * They have attributes such as column- and rownames to reference each entry in the table.

2. _‘Factors’_ (class `factor`) are object of type `integer` and hold categorical data.

    In statistics, many experiments involve the recording of categorical data, e.g. male and female, or the different cell lines, treatments etc. used in an experiment. As they would be stored as `character`, this can take a lot of memory to store.

    * Each category (‘level’) is assigned a number. Instead of the `character` vector, a much smaller `integer` vector is stored.
    * The assignments are saved as an attribute to the `factor` object.
    
    If you look for an example, here you go.
    
    ```{r}
x <- c("m", "f", "m", "m", "f", "m", "d")
x
typeof(x)
# create an object of class ‘factor’
x <- factor(x)
x
typeof(x)
# get the integer representation of the factor
as.integer(x)

# get the assignments of the levels
attributes(x)$levels
levels(x) # short-hand for objects of class ‘factor’
    ```

You can get the object's class with `class(x)` and the attributes associated with it with  `attributes(x)`.

> Having said that `print` is aware of an object's class, be reminded: The _good thing_ is that `print(x)` will show you only the ‘important’ aspects of the object. The _bad thing_ is that you can easily think that you are seeing the real object. In reality, you are just seeing the self-portrait of the object that it wants you to see.

## Vectorization in R

Recall that the basic unit in R is a vector. Sometimes there will be dozens, sometimes millions of elements stored in a single vector. Vectorization makes sure an operation treats the object as a whole rather than each element in the vector separately.

### Numerical Vectors and Operations

There are some convenience expressions to create regular sequences of numeric objects.

```{r}
# create a vector of type integer from 1 to 10 ...
x <- 1:10
# ... and backwards
x <- 10:1

typeof(x)

# create a vector from 1 to 5 (integer) ...
y <- seq(1, 5)
# ... spaced by 0.5 (double) ...
y <- seq(1, 5, by = 0.5)

typeof(y)

# concatenate both vectors
c(x, y)
```

> The numbers in squared brackets in the console output, `## [1]` and `## [15]`, are the indices of the following element, i.e. the second row of output starts with element number 15.

Note that the order of the elements in each vector is preserved upon concatenation.

***

Operations on a vector of length one are typically done element-by-element. 

```{r}
1:10 + 2
```

If the operation invovles two vectors of the same length, the operation is applied to each pair of elements.

```{r}
1:10 + 1:10
```

If the vectors are of different lengths, but one length is a multiple of the other, R resuses the shorter vector as needed.

```{r}
1:2 + 1:10
```

If the lengths are not a multiple of eachother, you will get a warning.

```{r}
1:10 + 1:3
```

### Logical Vectors and Operations

The operators `<` (less than), `>` (greater than), `<=` (less than or equal to), `>=` (greater than or equal to), `==` (equal to), and `!=` (not equal to) can be used to create logical vectors.

```{r}
x = 1:4
x > 2
```

We can assign the result of these expressions to a new variable.

```{r}
y <- x > 2
z <- x %% 2 == 1
```

The `!` sign is used to _negate_ logical vectors or boolean operations.

```{r}
!y
```

The boolean operators include `&` (and), `|` (or) and `xor(x, y)` (either or, but not both). They operate element-wise.

```{r}
y; z

y & z
y | z
xor(y, z)
```

To test if there is at least one logical `TRUE` in a vector, there is the `any(...)` function. Its complement is the `all(...)` function.

### Boolean Set Operations

Sometimes, you might need to know whether a value, e.g. `"c"`, is part of the following vectors or not.

```{r eval=FALSE}
vect_1 <- c("a", "b", "c")
vect_2 <- c("a", "b", "d")

# using basic logcial operators ...
any(vect_1 == "c")
any(vect_2 == "c")

# using set operation ...
"c" %in% vect_1
"c" %in% vect_2
```

Also, several fundamental operations to construct new sets from given sets are implemented in R.

```{r eval=FALSE}
# unions: elements of either set (without duplicates)
union(vect_1, vect_2)

# intersections: elements shared in both sets
intersect(vect_1, vect_2)

# complements: elements in one set, but not the other
setdiff(vect_1, vect_2)
setdiff(vect_2, vect_1)

# equality in membership
vect_3 <- c("b", "c", "a")
vect_3 == vect_1 # no
setequal(vect_3, vect_1) # yes
```

### Indexing Vectors

An index is used to refer to a specific element in a vector (or any other data structure). In R, square brackets are used to perfom indexing.

The usual form of indexing is `[`, which can select _more than one_ element.

```{r}
x <- 11:20
x[4] # fourth element in the vector
```

Numeric vectors can be used as index vectors.

```{r}
x[c(1, 3, 5)]
LETTERS[x]      # LETTERS is a built-in vector with A, B, C, ..., X, Y, Z
letters[x - 10] # letters is a built-in vector with a, b, c, ..., x, y, z
```

Negative values refer to indices to be excluded rather than included.

```{r}
x[c(-1, -3, -5)]
```

Logical vectors can be used as index vectors.

```{r}
x[x < 15]
x[!y]
```

Character vectors can be used as index vectors for named vectors. The advantage being that alphanumeric names are often easier to remember than numeric indices.

```{r}
noble_gases[c("Krypton", "Helium")]
```

`[[` can only be used to select a single element, dropping names.

```{r}
noble_gases[["Krypton"]]
```

### Indexing Lists

_(This section is included for sake of completeness. You may skip it.)_

In contrast to atomic vectors, lists can be recursively indexed with `[[`. Compare.

```{r}
z <- list(A = list(a = 1, b = 2), B = 3:5)
z

unlist(z) # can be used to flatten a list of lists

# returns a list of lists with one element that contains two lists, i.e. list z subsetted
z[1]
# returns a list with two elements, i.e. the first element of list z
z[[1]] 
# returns the second element of the first element of list z
z[[c(1, 2)]]
```

For named lists, the `$` operator can be used to browse the different levels in a list.

```{r}
# returns the second element of the first element of list z
z$A$b
# same as
z$"A"$"b"
# same as
z[["A"]][["b"]]
```

### Applying Functions to Vectors

The most powerful aspect of R is the vectorization of tasks. If you are familiar with some other programming language, this is equivalent to, but much faster than, writing a `for`-loop (which also exists in R).

Many functions in R are _vectorized_. If applied to a vector, the function will be applied to each of the elements in the whole vector ‘at once’. The result will be a vector.

```{r eval=FALSE}
# the command ...
sqrt(c(4, 9, 16, 25))
# ... has the same result as ...
c(sqrt(4), sqrt(9), sqrt(16), sqrt(25))
```

If we have more complex functions to apply on a vector or list, we can quickly make use of an `apply(...)` statement. There are six different kinds:

* `apply`  to go along in either direction of a matrix (not to be discussed in this workshop),
* `lapply`, `vapply` and `sapply` to go along vectors and lists, i.e. one-dimensional matrices; each of the three functions comes with a different set of options,
* `tapply` to split a vector or list based on a factor into groups and then apply a function to each group (useful, but not discussed, since we will have a different approach for `data.frame` objects), and
* `mapply` to go along multiple vectors or lists simultaneously (not discussed in this workshop). 

We will have a look on `lapply` only for the moment.

```{r}
x <- list(a = 1:11, b = 21:31, c = 100:200)
# calculate the mean for each element in x
lapply(x, mean)
```

We can also apply anonymous functions to the vector.

```{r}
# calculate the mean for each element i in x squared
lapply(x, function(i) mean(i**2))
```

Have a look how the one-liner from above would look as a `for`-loop.

```{r eval=FALSE}
for (i in x) {
  
  print(mean(i**2)) # no implicit printing from within flow control statetments
  
}
```

This would become even more tedious if we wanted to save the results as well in a vector.

```{r}
y <- x[] # copy attributes of x if any; in this case x has names

for (i in seq_along(x)) {   # need to create indexing variable manually
  
  y[[i]] <- mean(x[[i]]**2) # need to access each element individually

}

y
```

So, using `lapply` saves you a lot of work!

## Summary

In this introduction you have learned

* that there are different data types,
* that object classes are built from these ‘atomic’ data types,
* the basic unit in R is a vector (of one single data type),
* vectors can be indexed based on their value, their name or by position,
* vectorization of tasks applied to the elements of a vector is preferred over writing loops.

## Quick Questions

1. Which data type is the vector `c("Tagesgericht" = 1.60, "Menü 1" = 2.00, "Menü 2" = 2.50, "Aktionsteller" = 3.50)`?
2. Can you assign this vector to the variable `prizes_students`? How could you access the costs for “Menü 2”?
3. Which menus cost more than two euros?
4. Your colleagues order twice the “Tagesgericht”, three times “Menü 1”, three times “Menü 2”, and once “Aktionsteller”. The boss pays. How much does he or she need to pay?
5. The canteen wants to increase the prizes by 50 cents. Calculate the new prizes!
6. The prizes of the menus for employees are `c(3.10, 3.50, 4.00, 5.00)` respectively. Create a new vector called `prizes_employees` and assign the names of `prizes_students` to the names of this vector.
7. Let's make a list containing both vectors called `prizes`. If you want, you can include the prizes for guests, which are 50 cents higher than the costs for employees. Apply the rise in prices to all groups!
