==> part_13-summarizing_tables.Rmd <==
---
title: "Grouping and Summarizing Data"
output: 
  github_document:
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

***

You will learn in this section is how to summarize (large) data sets. Typically, the data is first split into several groups defined by a categorical variable (‘factor’) and then some sort of summary statistics are applied to each group.

If you have not done so yet, please attach the `tidyverse` and make sure you have the `plate_data` object from the [last section](part_12-manipulating_tables.html)--exercises included. If not, here you go.

```{r message=FALSE}
library(tidyverse)
readRDS("part_12-plate_data.rds") -> plate_data
```

## Making Groups

`group_by(...)` specifies the columns containing the categorical variables on which you want to split your data set. It is _not_ necessary (although sometimes useful) to convert the column into a `factor`.

Grouping does not change how the data looks like, apart from listing how it's grouped.

```{r}
plate_data %>% group_by(sample_id)
```

To check how many entries there are per group, there is `group_data()`.

```{r}
plate_data %>% group_by(sample_id, replicate_id) %>% group_data()
```

We could also use ‘scoped’ variants such as `group_by_at(...)`, `group_by_if(...)` or `group_by_all(...)` to select the columns on which to group.

```{r eval=FALSE}
plate_data %>% group_by_at(vars(ends_with("id"))) # group by "sample_id" and "replicate_id"
```

By default, `group_by(...)` overrides existing grouping of the data set. Use `group_by(..., add = TRUE)` to append.

To remove the grouping, use `ungroup()`.

## Summarizing Groups

A common summary of continuous data is to give the average and a measure of the spread e.g. using the standard deviation `sd(x)`, the median absolute deviation `mad(x)`, or the interquartile range `IQR(x)`. 

Let's have a look at a simple case first to identify the differences.

```{r}
df <- data.frame(g = rep(c("a", "b"), each = 2), x = seq(1, 4)) 

df %>% mutate(y = mean(x))
df %>% group_by(g) %>% mutate(y = mean(x))
df %>% group_by(g) %>% summarize(y = mean(x))
```

Given `plate_data`, let's calculate the average intensity for each `sample_id` and `concentration`.

```{r}
plate_data %>% 
  group_by(sample_id, concentration) %>% 
  # calculate mean and sd
  summarize(mean = mean(intensity, na.rm = TRUE), 
            sd = sd(intensity, na.rm = TRUE), 
            # count the rows that are not NA
            N  = sum(!is.na(intensity)))
```

Suspiciously, the standard deviation of some samples with four replicates are higher than the samples with just two replicates. Might there be something ‘wrong’ with one of `replicate_3` or `replicate_4`?

For categorical data, we are more likely to be interested in the number of observations, `n()`, or the number of unique values a variable takes within a group, `n_distinct(...)`. 

A routine task is for example to calculate the proportional composition.

```{r}
mtcars %>%
    group_by(am, gear) %>%
    summarise(n = n()) %>%
    mutate(freq = n / sum(n))

mtcars %>%
    group_by(am, gear) %>%
    summarise(n = n()) %>%
    mutate(freq = prop.table(n))
```


There are many more summary functions. Have a look on this [cheat sheet](https://github.com/rstudio/cheatsheets/raw/master/data-transformation.pdf).

## Comparing Groups

Statistics is a powerful way to quantify how different supposedly different groups are from each other. Let's see how to make uni- and multivariate comparisons or to fit models through data based on grouped `data.frame` objects in the `tidyverse`.

### Student's _t_ Test

In this section, we will use a built-in data set in R, `reshape2::tips`, which contains a record on the tips a waiter received over several months in a restaurant.

```{r}
reshape2::tips %>% as_tibble()
```

For example, we could be curious whether male or female guests give higher tips. In base R, we would go with this one-liner.

```{r}
t.test(tip ~ sex, data = reshape2::tips)
```

However, if we wanted to perform this test by weekday, we were left with subsetting the data by day and analyze each subset.

```{r}
reshape2::tips[which(reshape2::tips$day == "Fri"), ] -> tips.Fri
t.test(tip ~ sex, data = tips.Fri)
```
```{r include=FALSE}
rm(tips.Fri)
```

We would certainly prefer a neat table with all the results organized by the grouping variable. In a sense, a vectorized approach.

Let's look how to perform a Student's _t_ test within the `tidyverse`. 

We will use `group_modify(...)` to create a summary for cases when `summarize(...)` is too limited. `group_modify(...)` takes a function as argument, which operates on and returns back a `data.frame`. In this case, the function will be given the different `data.frame` objects, one for each group (here: Thu, Fri, Sat, Sun). 

> Since R's `t.test` does not return a `data.frame` by default, we wrap it with `~ broom::tidy(...)`. This function makes a `data.frame` containing the parameters (e.g. coefficients and slopes) of the model. 

> The `~` in front of `broom::tidy(...)` is mandatory for reasons beyond the scope of this tutorial at this point.

Within `group_modify`, we can refer to the current subset with `.` (or `.x`) and the value of the grouping variable with `.y`.

```{r}
reshape2::tips %>% 
  group_by(day) %>%
  group_modify(~ broom::tidy(t.test(tip ~ sex, data = .)))
```

A little down-side of this nicely general approach is that column names may become rather enigmatic. This is because they need to apply to very different cases in which `broom::tidy(...)` is used. 

Here, the columns named `estimate`, `estimate1` and `estimate2` refer to the mean difference between both groups, the mean of the first group (here: ‘Female’) and the mean of the second group (here: ‘Male’) respectively.

### Student's _t_ Test with Multiple Comparisons

If more than two groups are compared, we should adjust the p-values for the additional comparisons. In R, the function is a little bit misleadingly named `pairwise.t.test(x, g)`, but can be used for paired and unpaired comparisons between multiple groups.

First, the same as above.

```{r}
reshape2::tips %>% 
  group_by(day) %>% 
  group_modify(~ broom::tidy(pairwise.t.test(
    x = .$tip, # the column with the values to compare
    g = .$sex, # the column from which the different groups are taken
    p.adjust.method = "none", pool.sd = FALSE)))
```

Let's group by sex and compare the tips for each weekday without applying a correction.

```{r}
reshape2::tips %>% 
  group_by(sex) %>% 
  group_modify(~ broom::tidy(pairwise.t.test(
    x = .$tip, g = .$day, p.adjust.method = "none", pool.sd = FALSE)))
```

However, when properly done, the ‘significant’ difference for female payers between Thursdays and Sundays vanishes.

```{r eval=FALSE}
reshape2::tips %>% 
  group_by(sex) %>% 
  group_modify(~ broom::tidy(pairwise.t.test(
    x = .$tip, g = .$day, p.adjust.method = "hochberg", pool.sd = FALSE)))
```

This approach can also be applied to other hypothesis testing functions such as `cor.test`, `wilcox.test`, `chisq.test` etc.

### Linear Regression and Beyond

Another example on the same lines is a linear regression by group. Let's try to describe the `intensity` observed in `plate_data` as a function of the `concentration`. Such a model is fit in R with the linear model `lm` call.

```{r}
plate_data %>% 
  group_by(sample_id) %>% 
  group_modify(~ broom::tidy(lm(intensity ~ log10(concentration + 1), data = .)))
```

So far, we have used `broom::tidy` to get the parameters of the model. If we are interested in the statistical summary of the overall model, we use `broom::glance`.

```{r}
plate_data %>% 
  group_by(sample_id) %>% 
  group_modify(~ broom::glance(lm(intensity ~ log10(concentration + 1), data = .)))
```

All `broom` functions apply equally well to the output of generalized linear and non-linear models.

## Hands-On Exercise

Unfortunately, some of the fits on `plate_data` are not of satisfactory quality. Maybe we can figure out if this is because of a bad replicate in our data.

1. We will group `plate_data` by `sample_id` and `replicate_id`. One of the groups has no values recorded at all, which will cause troubles in later steps. Figure out which one it is and exlude it by filtering the rows accordingly.

```{r include=FALSE}
# solution 1: remove all NA cases from the begin
plate_data %>% 
  filter(!is.na(intensity))

# solution 2: remove groups in which all values are NA
plate_data %>% 
  group_by(sample_id, replicate_id) %>% 
  filter(!all(is.na(intensity)))
```

2. Now, perform the fit with `lm` and have a look only at the slopes. How much do they differ between replicates?

```{r include=FALSE}
# solution 1: manually compare differences
plate_data %>% 
  group_by(sample_id, replicate_id) %>% 
  filter(!all(is.na(intensity))) %>% 
  group_modify(~ broom::tidy(lm(intensity ~ log10(concentration + 1), data = .))) %>% 
  filter(term != "(Intercept)")

# solution 2: let's use R
plate_data %>% 
  group_by(sample_id, replicate_id) %>% 
  filter(!all(is.na(intensity))) %>% 
  group_modify(~ broom::tidy(lm(intensity ~ log10(concentration + 1), data = .))) %>% 
  filter(term != "(Intercept)") %>% 
  # within each sample, one of the replicates must have the highest deviance from the
  # mean estimate ...
  group_by(sample_id) %>% 
  mutate(estimate_diff = abs(estimate - mean(estimate))) %>% 
  top_n(1, estimate_diff) %>% select(sample_id, replicate_id, estimate, estimate_diff)
```
