==> part_23-plot_themes.Rmd <==
---
title: "Changing the Appearance of a Plot"
output: 
  github_document:
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.width = 6, fig.height = 3, dpi = 120, warning = FALSE)
```

***

Customization of plots is a vast field, so this section cannot be but a brief glimpse into what is possible.

```{r message=FALSE}
library(tidyverse)
```

## The Fabric and Flowers to Choose From

### Color Schemes

First, some frequently encountered tips with regards to colors.

* Avoid unnecessary use of color.
* Be consistent in the use of color.
* Ideally use color-blind friendly colors.

> Independent of your software, screen colours and printed colours will vary!

In R, color can be represented by 

* a hexadecimal constant, e.g. `"#faebd7"`, 
* a function that transforms mixing ratios into such a constant, e.g. `rgb(250, 235, 215)` or `hcl(...)` or `hsv(...)`, or
* a color name (or its index) in the `colors()` vector, e.g. `"antiquewhite"`.

There are `r length(colors())` colors in R! Make your choice.

```{r}
colors() %>% head(100)
```

Anything can be colored. Here are some ways to change the fill color of the bar graphs according to `sample_id`.

```{r}
mpg %>%
  filter(manufacturer %in% unique(manufacturer)[1:5]) %>% 
  ggplot(aes(x = manufacturer, y = hwy, fill = manufacturer)) +
  geom_bar(stat = "summary", fun.y = mean) +
  geom_errorbar(stat = "summary", fun.data = mean_se, width = .2) ->
  p

# with self-defined color scale
p + scale_fill_manual(values = c(
  audi      = "magenta", 
  chevrolet = "#dd3497", 
  dodge     = rgb(174, 1, 126, maxColorValue = 255), 
  ford      = rgb(0.5, 0, 0.5), 
  honda     = hcl(0, 35, 85),
  whatever  = "green"
))
```
```{r eval=FALSE}
# with pre-defined color scales

p + scale_fill_grey()      # grey scale
p + scale_fill_viridis_d() # color blind and color printing friendly palette

# with pre-defined color scales from RColorBrewer

p + scale_fill_brewer(palette = "Pastel1")
```
```{r fig.width=5, fig.height=9}
RColorBrewer::display.brewer.all()
```

### Stripes and Patterns

These are disencouraged and therefore not supported in `ggplot2`.

### Point Shapes

Typically, the color of the entire shape is set with `color = ...`. Only some shapes accept a different `fill = ...` color from the border color.

```{r echo=FALSE, fig.height=2}
data.frame(p = 0:25) %>%
  ggplot() +
  scale_y_reverse(name = "", expand = expand_scale(add = .5)) +
  scale_x_continuous(name = "") +
  scale_shape_identity() +
  geom_point(mapping = aes(x = p %% 13, y = p %/% 13 + .2, shape = p), size = 5, fill = "magenta") +
  geom_text(mapping = aes(x = p %% 13, y = p %/% 13 + .6, label = p), size = 5)+
  theme_minimal() + theme(
    panel.grid = element_blank(),
    axis.text  = element_blank(),
    axis.title = element_blank()
  )
```

### Linetypes

```{r echo=FALSE, fig.height=2, fig.width=2.5}
data.frame(p = 0:6, t = c("blank", "solid", "dashed", "dotted", "dotdash", "longdash", "twodash")) %>%
  ggplot() +
  scale_y_reverse(name = "", expand = expand_scale(add = .5)) +
  scale_x_continuous(name = "") +
  scale_linetype_identity() +
  geom_segment(mapping = aes(y = p, yend = p, x = .05, xend = 1, linetype = p)) +
  geom_text(mapping = aes(x = 0, y = p, label = p), size = 5, hjust = 1) +
  geom_text(mapping = aes(x = 1.05, y = p, label = str_c('"', t, '"')), hjust = 0) +
  geom_vline(xintercept = 1.5, linetype = 0) +
  theme_minimal() + theme(
    panel.grid = element_blank(),
    axis.text  = element_blank(),
    axis.title = element_blank()
  )
```

To modify the line end style, there is the parameter `lineend = ...`, which can take the values `"round"`, `"butt"`, or `"square"`.

## Common Dimensions and Layout Parameters

| parameter | effect |
|-----------|--------|
| `unit` | unit of dimensions; defaults `"pt"` |
| `size` | line/border size in `"mm"`; text size in `"pt"` |

This means, if you want to specify the line width in points, you will need the following conversion.

```{r eval=FALSE}
size = grid::convertUnit(unit(8, "pt"), "mm", valueOnly = TRUE)
```

| parameter | effect |
|-----------|--------|
| `hjust` | horizontal anchor between 0 (left) and 1 (right) |
| `vjust` | vertical anchor between 0 (bottom) and 1 (top) |
| `angle` | rotation around justification anchor |

An overview of the various effects; modified from [this source](https://gist.github.com/cdesante/4070548).

```{r echo=FALSE, fig.height=3}
expand.grid(
  hjust=c(0, 0.5, 1),
  vjust=c(0, 0.5, 1),
  angle=c(0, 45, 90),
  text="is-R"
) %>% ggplot(aes(x=hjust, y=vjust)) + 
  geom_point() +
  geom_text(aes(label=text, angle=angle, hjust=hjust, vjust=vjust)) + 
  scale_x_continuous(breaks=c(0, 0.5, 1), expand=c(0, 0.2)) +
  scale_y_continuous(breaks=c(0, 0.5, 1), expand=c(0, 0.2)) + 
  coord_fixed() +
  facet_grid(cols = vars(angle), labeller = labeller(angle = function(x) str_c("angle = ", x))) +
  theme_bw() + theme(
    strip.text = element_text(size = 11),
    axis.title = element_text(size = 11),
    axis.text  = element_text(size = 11),
    panel.grid.minor = element_blank()
  )
```

## Themes

A `theme` controls all non-data parts of a plot. This is anything from the default shading of the canvas, the line widths of the grid or the color of the axes labels.

### Predefined Themes

Let's briefly look how the default themes in `ggplot2` look like. 

```{r}
p <- p + facet_wrap(vars(year))
p + theme_gray() # the default theme in ggplot2

p + theme_bw()
p + theme_linedraw()
p + theme_light()
p + theme_dark()

p + theme_classic()

p + theme_void() # a completely empty theme
```

### Modifying Components of a Theme

Parameters to change the style or layout of distinct elements in a plot are set via `theme(...)`. 

```{r}
p + scale_fill_grey() + theme_grey() + theme(
  # place legend below plot
  legend.position = "bottom",
  # remove major gird lines on x axis
  panel.grid.major.x = element_blank(),
  # color minor grid lines on y axis
  panel.grid.minor.y = element_line(color = "red", linetype = "dashed"),
  # fill facet labels
  strip.background = element_rect(fill = "yellow"),
  # make text on y axis larger
  axis.text.y = element_text(size = 14),
  # highlight legend
  legend.box.background = element_rect(color = "green", size = 1)
)
```

Some frequently and less frequently used parameters are given below.

| parameter | modifies |
|-----------|----------|
| <b>general</b> | these modify _all_ occurences |
| `line` | all line elements with `element_line()` |
| `rect` | all rectangular elements with `element_rect()` |
| `text` | all text elements with `element_text()` |
| `title` | all title elements (plot, axes, legends) with `element_text()` |
| <b>legend</b> |
| `legend.position` | position of the legend, e.g. `"none"`, `"left"`, `"top"` etc. |
| `legend.background` | background of legend |
| `legend.title` |
| `legend.title.align` |
| `legend.text` |
| `legend.text.align` |
| `legend.key` | background underneath keys, via `element_rect()` |
| `legend.key.size` | size via `unit(...)` |
| `legend.direction` | layout of items, either `"horizontal"` or `"vertical"` |
| `legend.justification` | anchor point for positioning legend inside plot |
| `legend.box` | layout of multiple legends, either `"horizontal"` or `"vertical"` |
| <b>axes titles</b> | 
| `axis.title` | both |
| `axis.title.x` | x axis, e.g. of `xlab(...)` |
| `axis.title.y` | y axis, e.g. of `ylab(...)` |
| <b>axes texts</b> |
| `axis.text` | all |
| `axis.text.x` | x axis, e.g. of `scale_x_log10(label = ...)` |
| `axis.text.y` | y axis, e.g. of `scale_y_log10(label = ...)` |
| <b>axes ticks and lines</b> |
| `axis.ticks` | style via `element_line()` |
| `axis.ticks.x` | x axis |
| `axis.ticks.y` | y axis |
| `axis.ticks.length` | length via `unit(...)` |
| `axis.ticks.length.x` | x axis |
| `axis.ticks.length.y` | y axis |
| `axis.line` | style via `element_line()` |
| `axis.line.x` | x axis |
| `axis.line.y` | y axis |
| <b>canvas underneath data</b> |
| `panel.background` |
| `panel.border` |
| `panel.spacing` | between facet panels; length via `unit(...)` |
| `panel.grid` | style via `element_line()` |
| `panel.grid.major` |
| `panel.grid.minor` |
| `panel.grid.major.x` |
| `panel.grid.major.y` |
| `panel.grid.minor.x` |
| `panel.grid.minor.y` |
| <b>facet labels</b> |
| `strip.background` |
| `strip.background.x` |
| `strip.background.y` |
| `strip.text` |
| `strip.text.x` |
| `strip.text.y` |
| <b>canvas around data</b> |
| `plot.background` | drawn underneath the entire plot |
| `plot.title` |
| `plot.subtitle` |
| `plot.caption` |
| `plot.tag` |
| `plot.tag.position` |
| `plot.margin` | around the entire plot; length via `unit(...)` |

<!-- ### Creating Your Own Theme -->
